js & waypoints
Added my js variable for using waypoints for the svg's in the contact
section. Put it so that the default setting for the svg's have no
animation, and put the animations on the js-active. The svg's do not
start animating until all of the icons are showing at the same time.